feat(health_record): 건강 기록 생성 및 조회 API 구현

+++ src/main/java/io/github/jo0yo0n/vitalsjournal/threshold/controller/response/ThresholdResponse.java
@@ -2,11 +2,10 @@ package io.github.jo0yo0n.vitalsjournal.threshold.controller.response;
 
 import io.github.jo0yo0n.vitalsjournal.threshold.domain.Threshold;
 import io.github.jo0yo0n.vitalsjournal.threshold.domain.ThresholdMetric;
-import java.math.BigDecimal;
 import java.time.Instant;
 
 public record ThresholdResponse(
-    Long id, ThresholdMetric metric, BigDecimal minValue, BigDecimal maxValue, Instant updatedAt) {
+    Long id, ThresholdMetric metric, Short minValue, Short maxValue, Instant updatedAt) {
 
   public static ThresholdResponse from(Threshold threshold) {
     return new ThresholdResponse(
